requirements.txt
pyyaml>=6.0
python-dotenv>=0.19.0
openai>=1.0.0
anthropic>=0.3.0
cohere>=4.0.0
typing-extensions>=4.0.0
aiohttp>=3.8.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
responses>=0.23.0  # For mocking HTTP requests
